% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adhoc3}
\alias{adhoc3}
\title{TREC-3 Ad hoc track.}
\format{A data frame with 40 columns (systems) and 50 rows (queries).}
\usage{
adhoc3
}
\description{
This is the set of Average Precision scores of the 40 systems submitted to the TREC-3 Ad hoc
track, evaluated over 50 topics.
}
\references{
D. Harman (1994). Overview of the Third Text REtrieval Conference (TREC-3). Text
  REtrieval Conference.
}
\seealso{
\url{http://trec.nist.gov}
}
\keyword{datasets}
