% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic4}
\alias{synthetic4}
\title{Synthetic dataset no. 4.}
\format{A data frame with 10 columns (systems) and 12 rows (queries).}
\usage{
synthetic4
}
\description{
This is the Synthetic dataset no. 4 from Table 3.2 on page 73 of Brennan (2001), recasted as a p
x i design, as required on page 182.
}
\references{
R.L. Brennan, "Generalizability Theory". Springer, 2001.
}
\keyword{datasets}
